Always return decorated queries
